+++ fixture/change.js
@@ -1,0 +1,85 @@
+const TC = [
+  {
+    input: '40660',
+    change: {
+      500: 81,
+      100: 1,
+      50: 1,
+      10: 1,
+    },
+  },
+  {
+    input: 10,
+    change: {
+      500: 0,
+      100: 0,
+      50: 0,
+      10: 1,
+    },
+  },
+  {
+    input: 50,
+    change: {
+      500: 0,
+      100: 0,
+      50: 1,
+      10: 0,
+    },
+  },
+  {
+    input: '1000',
+    change: {
+      500: 2,
+      100: 0,
+      50: 0,
+      10: 0,
+    },
+  },
+  {
+    input: 20,
+    change: {
+      500: 0,
+      100: 0,
+      50: 0,
+      10: 2,
+    },
+  },
+  {
+    input: '600',
+    change: {
+      500: 1,
+      100: 1,
+      50: 0,
+      10: 0,
+    },
+  },
+  {
+    input: '650',
+    change: {
+      500: 1,
+      100: 1,
+      50: 1,
+      10: 0,
+    },
+  },
+  {
+    input: '11',
+    change: {
+      500: 0,
+      100: 0,
+      50: 0,
+      10: 1,
+    },
+  },
+  {
+    input: '659',
+    change: {
+      500: 1,
+      100: 1,
+      50: 1,
+      10: 0,
+    },
+  },
+];
+
+export default TC;
