+++ src/js/model/VendingMachine.test.js
@@ -1,6 +1,11 @@
 /* eslint-disable max-lines-per-function */
 import { JSDOM } from 'jsdom';
 import VendingMachine from './VendingMachine';
+import CHANGE_TC from '../../../fixture/change';
+import {
+  TC_ACC as CHARGE_TC_ACC,
+  TC_NONE_ACC as CHARGE_TC__NONE_ACC,
+} from '../../../fixture/charge';
 
 const { window } = new JSDOM();
 global.window = window;
@@ -15,34 +20,7 @@ describe('금액 충전', () => {
   });
 
   it('최소 충전 금액은 10원 이상, 50000원 이하이다.', () => {
-    const TC = [
-      {
-        input: '1',
-        answer: 0,
-      },
-      {
-        input: '9',
-        answer: 0,
-      },
-      {
-        input: 10,
-        answer: 10,
-      },
-      {
-        input: '30000',
-        answer: 30000,
-      },
-      {
-        input: '50000',
-        answer: 50000,
-      },
-      {
-        input: '50001',
-        answer: 0,
-      },
-    ];
-
-    TC.forEach(({ input, answer }) => {
+    CHARGE_TC__NONE_ACC.forEach(({ input, answer }) => {
       vendingMachine = new VendingMachine();
 
       vendingMachine.input(input);
@@ -52,34 +30,7 @@ describe('금액 충전', () => {
   });
 
   it('사용자는 금액 충전을 누적으로 할 수 있다.', () => {
-    const TC = [
-      {
-        input: '5',
-        answer: 0,
-      },
-      {
-        input: '5',
-        answer: 0,
-      },
-      {
-        input: 10000,
-        answer: 10000,
-      },
-      {
-        input: '30000',
-        answer: 40000,
-      },
-      {
-        input: '10000',
-        answer: 50000,
-      },
-      {
-        input: '50001',
-        answer: 50000,
-      },
-    ];
-
-    TC.forEach(({ input, answer }) => {
+    CHARGE_TC_ACC.forEach(({ input, answer }) => {
       vendingMachine.input(input);
 
       expect(Number(vendingMachine.chargedMoney)).toEqual(answer);
@@ -125,91 +76,7 @@ describe('잔돈 반환', () => {
   });
 
   it('잔돈은 최소 개수의 동전으로 반환된다.', () => {
-    const TC = [
-      {
-        input: '40660',
-        change: {
-          500: 81,
-          100: 1,
-          50: 1,
-          10: 1,
-        },
-      },
-      {
-        input: 10,
-        change: {
-          500: 0,
-          100: 0,
-          50: 0,
-          10: 1,
-        },
-      },
-      {
-        input: 50,
-        change: {
-          500: 0,
-          100: 0,
-          50: 1,
-          10: 0,
-        },
-      },
-      {
-        input: '1000',
-        change: {
-          500: 2,
-          100: 0,
-          50: 0,
-          10: 0,
-        },
-      },
-      {
-        input: 20,
-        change: {
-          500: 0,
-          100: 0,
-          50: 0,
-          10: 2,
-        },
-      },
-      {
-        input: '600',
-        change: {
-          500: 1,
-          100: 1,
-          50: 0,
-          10: 0,
-        },
-      },
-      {
-        input: '650',
-        change: {
-          500: 1,
-          100: 1,
-          50: 1,
-          10: 0,
-        },
-      },
-      {
-        input: '11',
-        change: {
-          500: 0,
-          100: 0,
-          50: 0,
-          10: 1,
-        },
-      },
-      {
-        input: '659',
-        change: {
-          500: 1,
-          100: 1,
-          50: 1,
-          10: 0,
-        },
-      },
-    ];
-
-    TC.forEach(({ input, change }) => {
+    CHANGE_TC.forEach(({ input, change }) => {
       vendingMachine.input(input);
       const result = vendingMachine.change();
